requirements.txt
flask
flask-sqlalchemy
mysql-connector-python
pyyaml
python-dotenv